Consolidate recovery headers and implement dual robot system control on the real DRDK API
