+++ flexiv_hardware/include/flexiv_hardware/recovery_node.hpp
@@ -23,7 +23,7 @@
 #include "flexiv_msgs/msg/robot_event.hpp"
 #include "flexiv_msgs/srv/get_operational_status.hpp"
 
-#include "flexiv_hardware/driver_status.hpp"
+#include "flexiv_hardware/fault_recovery.hpp"
 #include "flexiv_hardware/robot_system_control.hpp"
 
 namespace flexiv_hardware {
